# Minor deserialisation optimisations