--- a/Agda.cabal
+++ b/Agda.cabal
@@ -244,6 +244,7 @@ library
                 , regex-tdfa >= 1.3.1.0 && < 1.4
                 , split >= 0.2.0.0 && < 0.2.4
                 , stm >= 2.4.4 && < 2.6
+                , STMonadTrans >= 0.4.3 && < 0.5
                 , strict >= 0.3.2 && < 0.5
                 , template-haskell >= 2.11.0.0 && < 2.19
                 , text >= 1.2.3.0 && < 2.1