@SquidDev 2022-06-30
I'm able to eck out another 5% by writing to an array directly, rather than going through a list (bringing the stdlib down to 31.4s and agda-categories to 16.4s). However I make liberal use of `unsafePerformIO` - I don't know if this is a change people would be comfortable integrating? I know for one I'm never confident writing code which relies on it!

```haskell
instance B.Binary a => B.Binary (ListLike a) where
  put = __IMPOSSIBLE__ -- Will never deserialise this
  get = do
    n <- B.get :: B.Get Int
    arr <- return (unsafePerformIO $ newArray_ (0, fromIntegral n - 1) :: IOArray Int32 a)

    let
      getMany i = if i == n then return () else do
        x <- B.get
        () <- return $ unsafePerformIO $ unsafeWrite arr i x
        getMany (i + 1)
    () <- getMany 0

    arr <- return $ unsafePerformIO $ unsafeFreeze arr
    return $! ListLike arr
```