@andreasabel 2022-07-01
Great work, @SquidDev !

Concerning `unsafePerformIO`,  I share your confidence issues.  Alternatives:
- Could one instead replace the `Binary instance` for `Listlike` by some method living in `IO` and use it as such from `decode` (which lives in `TCM` and has `IO`)?  
- Give precise reasoning why this pattern of using `unsafePerformIO` is safe (also when used concurrently).
- Skip the extra 5% (or is it rather 7%) improvement.