@nad 2022-07-01
Perhaps one could use the `STT` monad transformer (see [Control.Monad.ST.Trans](https://hackage.haskell.org/package/STMonadTrans-0.4.6/docs/Control-Monad-ST-Trans.html)).