@SquidDev 2022-07-01
>  Could one instead replace the `Binary instance` for `Listlike` by some method living in `IO` and use it as such from `decode` (which lives in `TCM` and has `IO`)?

I think the main problem here is that we'd end up needing to implement most of `Binary` to do this, which is obviously unfortunate.

> Give precise reasoning why this pattern of using `unsafePerformIO` is safe (also when used concurrently).

Looking at the optimised core, I can kinda convince myself. But that's obviously only for one version of GHC. I think the problem is it breaks so many rules of "how to use unsafePerformIO" it's hard to guarantee anything.

> Perhaps one could use the `STT` monad transformer (see [Control.Monad.ST.Trans](https://hackage.haskell.org/package/STMonadTrans-0.4.6/docs/Control-Monad-ST-Trans.html)).

Was actually just looking at that! It appears to work, and (unsurprisingly) has the same performance as the unsafe `IO` version, so I it's worth switching over to that if people are happy to add it as a dependency?

> Skip the extra 5% (or is it rather 7%) improvement.

Should probably clarify here - it _is_ closer to 6%, but results are noisy (±1%, my machine is not set up for benchmarking :)), so erring on the more conservative side.