@nad 2022-07-01
> Was actually just looking at that! It appears to work, and (unsurprisingly) has the same performance as the unsafe `IO` version, so I think it's worth switching over to that if people are happy to add it as a dependency?

Apparently Andreas is already maintaining this package, so I think it is fine to add it as a dependency.