@SquidDev 2022-07-01
There we go! Numbers for the curious:

<details><summary>agda-stdlib</summary>

```
Total                      31,345ms           
Miscellaneous                   4ms           
Deserialization            19,621ms (30,964ms)
Deserialization.Compaction 11,342ms           
Import                        301ms           
Parsing                        75ms           
  32,712,768,384 bytes allocated in the heap
  12,955,373,760 bytes copied during GC
   1,217,073,128 bytes maximum residency (39 sample(s))
       1,523,736 bytes maximum slop
            1809 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     28373 colls,     0 par   12.642s  12.768s     0.0004s    0.0271s
  Gen  1        39 colls,     0 par    0.943s   0.943s     0.0242s    0.1166s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time   17.818s  ( 17.591s elapsed)
  GC      time   13.585s  ( 13.711s elapsed)
  EXIT    time    0.002s  (  0.007s elapsed)
  Total   time   31.406s  ( 31.311s elapsed)

  Alloc rate    1,835,899,498 bytes per MUT second

  Productivity  56.7% of total user, 56.2% of total elapsed
```

</details>

<details><summary>agda-categories</summary>

```
Total                      16,236ms           
Miscellaneous                   5ms           
Deserialization            10,179ms (16,039ms)
Deserialization.Compaction  5,860ms           
Import                        174ms           
Parsing                        16ms           
  21,730,347,856 bytes allocated in the heap
     844,509,472 bytes copied during GC
     794,239,640 bytes maximum residency (8 sample(s))
       1,956,608 bytes maximum slop
             990 MiB total memory in use (33 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       139 colls,     0 par    1.369s   1.371s     0.0099s    0.0562s
  Gen  1         8 colls,     0 par    0.164s   0.164s     0.0205s    0.0537s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.003s  (  0.003s elapsed)
  MUT     time   14.726s  ( 14.696s elapsed)
  GC      time    1.534s  (  1.535s elapsed)
  EXIT    time    0.001s  (  0.007s elapsed)
  Total   time   16.264s  ( 16.241s elapsed)

  Alloc rate    1,475,648,977 bytes per MUT second

  Productivity  90.5% of total user, 90.5% of total elapsed
```

</details>